+++ radicle-surf/src/git/namespace.rs
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::vcs::git::error::Error;
+use crate::git::error::Error;
 use nonempty::NonEmpty;
 pub use radicle_git_ext::Oid;
 use std::{convert::TryFrom, fmt, str};
